ARM: decompressor: Enable unaligned memory access for v6 and above
Modern GCC can generate code which makes use of the CPU's native
unaligned memory access capabilities.  This is useful for the C
decompressor implementations used for unpacking compressed kernels.

This patch disables alignment faults and enables the v6 unaligned
access model on CPUs which support these features (i.e., v6 and
later), allowing full unaligned access support for C code in the
decompressor.

The decompressor C code must not be built to assume that unaligned
access works if support for v5 or older platforms is included in
the kernel.

For correct code generation, C decompressor code must always use
the get_unaligned and put_unaligned accessors when dealing with
unaligned pointers, regardless of this patch.

Signed-off-by: Dave Martin <dave.martin@linaro.org>
Acked-by: Nicolas Pitre <nico@linaro.org>
---
This is the same as the previous post, with an additional comment
in the commit message regarding the use of {get,put}_unaligned,
as suggested by Nico.

Tested on ARM1136JF-S (Integrator/CP) and ARM1176JZF-S (RealView
PB1176JZF-S).  ARM1176 is like v7 regarding the MIDR and SCTLR
alignment control bits, so this tests the v7 code path.

link:https://patches.linaro.org/patch/12680/