fbcon: queue work on power efficient wq
fbcon uses workqueues and it has no real dependency of scheduling these
on the cpu which scheduled them.

On a idle system, it is observed that and idle cpu wakes up many times
just to service this work. It would be better if we can schedule it on a
cpu which the scheduler believes to be the most appropriate one.

This patch replaces system_wq with system_power_efficient_wq.

Cc: Dave Airlie <davem@davemloft.net>
Cc: netdev@vger.kernel.org
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>